face and sound

+++ src/GameObject.cpp
@@ -16,13 +16,13 @@ GameObject::~GameObject () {
 }
 
 void GameObject::Update (float dt) {
-    for (int i=0; i<components.size(); i++) {
+    for (int i=0; i < (int)components.size(); i++) {
         components[i]->Update(dt);
     }
 }
 
 void GameObject::Render () {
-    for (int i=0; i<components.size(); i++) {
+    for (int i=0; i < (int)components.size(); i++) {
         components[i]->Render();
     }
 }
@@ -48,7 +48,7 @@ void GameObject::RemoveComponent (Component* cpt) {
 }
 
 Component* GameObject::GetComponent (std::string type) {
-    for (int i=0; i<components.size(); i++) {
+    for (int i=0; i < (int)components.size(); i++) {
         if (components[i]->Is(type)) {
             return components[i];
         }
